Deduplicate Dav1dContext
